Organized and Refactored code for easier understanding
